Primitives should not be boxed just for "String" conversion
fix other occurrences

rebase

Return of boolean expressions should not be wrapped into an "if-then-else" statement

The operator ("<>") should be used.

fix files after merging branches

rebase-1

Return of boolean expressions should not be wrapped into an "if-then-else" statement

The operator ("<>") should be used.

fix reverted files after rebase

Return of boolean expressions should not be wrapped into an "if-then-else" statement

The operator ("<>") should be used.